Move NMR Firework to separate module
