+++ atomate/vasp/fireworks/__init__.py
@@ -1,3 +1,4 @@
 from __future__ import unicode_literals
 
 from .core import *
+from .nmr import *
